requirements.txt
Flask
Flask-JWT-Extended
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-SQLAlchemy
requests
psycopg2-binary
python-dotenv
gunicorn
marshmallow
SQLAlchemy
